Feat/test player and health bar

+++ Assets/Scripts/Interface/IDamagable.cs
@@ -4,5 +4,5 @@ using UnityEngine;
 
 public interface IDamagable
 {
-    public void TakeDamage(int amount);
+    public void TakeDamage(float amount);
 }
